event ccalendar event detail screen

+++ lib/services/fire_storage.dart
@@ -38,11 +38,11 @@ class FireStorage {
   }
 
   Future<void> deleteUserStorageInfo() async {
-    //TODO: Info is not deleted because the auth object is null
     try {
-      final userId = _user!.uid;
-
-      await _fireStore.ref('images/profile_images/$userId').delete();
+      final userId = _user?.uid;
+      if (userId != null) {
+        await _fireStore.ref('images/profile_images/$userId').delete();
+      }
     } on FirebaseStorage catch (e) {
       debugPrint(e.toString());
       rethrow;
